chore: change emodeId to uint8

+++ src/AaveV3Avalanche.sol
@@ -358,9 +358,9 @@ library AaveV3AvalancheAssets {
 }
 
 library AaveV3AvalancheEModes {
-  uint256 internal constant NONE = 0;
+  uint8 internal constant NONE = 0;
 
-  uint256 internal constant STABLECOINS = 1;
+  uint8 internal constant STABLECOINS = 1;
 
-  uint256 internal constant AVAX_CORRELATED = 2;
+  uint8 internal constant AVAX_CORRELATED = 2;
 }
