+++ src/AaveV3Ethereum.sol
@@ -623,7 +623,7 @@ library AaveV3EthereumAssets {
 }
 
 library AaveV3EthereumEModes {
-  uint256 internal constant NONE = 0;
+  uint8 internal constant NONE = 0;
 
-  uint256 internal constant ETH_CORRELATED = 1;
+  uint8 internal constant ETH_CORRELATED = 1;
 }
