+++ src/AaveV3Fuji.sol
@@ -241,7 +241,7 @@ library AaveV3FujiAssets {
 }
 
 library AaveV3FujiEModes {
-  uint256 internal constant NONE = 0;
+  uint8 internal constant NONE = 0;
 
-  uint256 internal constant STABLECOINS = 1;
+  uint8 internal constant STABLECOINS = 1;
 }
